recast redundantly throttle time variable as int

+++ src/main/java/Main.java
@@ -82,7 +82,7 @@ public class Main {
 	  // MAX_VERSION --- INT16, 2bytes, short --- 4
 	  outputBuffer.putShort((short) 4);
 	  // Throttle time --- INT32, 4 bytes --- 0
-	  outputBuffer.putInt(0);
+	  outputBuffer.putInt((int) 0);
 	  // TAGGED_FIELDS --- 1 byte --- now 0 'no tagged fields' --- not sure if MAX INT64
 	  outputBuffer.put((byte) 0);
